move generation code to generate module
